DataGraph index calculation refactored (moved to collector)

+++ MAVGCL/src/com/comino/flight/widgets/charts/control/IChartControl.java
@@ -18,13 +18,14 @@ package com.comino.flight.widgets.charts.control;
 
 import javafx.beans.property.BooleanProperty;
 import javafx.beans.property.DoubleProperty;
+import javafx.beans.property.FloatProperty;
 import javafx.beans.property.IntegerProperty;
 
 public interface IChartControl {
 
 	public BooleanProperty getCollectingProperty();
 	public IntegerProperty getTimeFrameProperty();
-	public IntegerProperty  getScrollProperty();
+	public FloatProperty  getScrollProperty();
 
 	public boolean isVisible();
